% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CADMUN}
\alias{CADMUN}
\title{City Structure}
\format{
A data frame with 5652 rows and 28 variables:
\describe{
  \item{\code{MUNCOD}}{integer COLUMN_DESCRIPTION}
  \item{\code{MUNCODDV}}{integer COLUMN_DESCRIPTION}
  \item{\code{SITUACAO}}{integer COLUMN_DESCRIPTION}
  \item{\code{MUNSINP}}{integer COLUMN_DESCRIPTION}
  \item{\code{MUNSIAFI}}{integer COLUMN_DESCRIPTION}
  \item{\code{MUNNOME}}{integer COLUMN_DESCRIPTION}
  \item{\code{MUNNOMEX}}{integer COLUMN_DESCRIPTION}
  \item{\code{OBSERV}}{integer COLUMN_DESCRIPTION}
  \item{\code{MUNSINON}}{integer COLUMN_DESCRIPTION}
  \item{\code{MUNSINONDV}}{integer COLUMN_DESCRIPTION}
  \item{\code{AMAZONIA}}{integer COLUMN_DESCRIPTION}
  \item{\code{FRONTEIRA}}{integer COLUMN_DESCRIPTION}
  \item{\code{CAPITAL}}{integer COLUMN_DESCRIPTION}
  \item{\code{UFCOD}}{integer COLUMN_DESCRIPTION}
  \item{\code{MESOCOD}}{integer COLUMN_DESCRIPTION}
  \item{\code{MICROCOD}}{integer COLUMN_DESCRIPTION}
  \item{\code{MSAUDCOD}}{integer COLUMN_DESCRIPTION}
  \item{\code{RSAUDCOD}}{integer COLUMN_DESCRIPTION}
  \item{\code{CSAUDCOD}}{integer COLUMN_DESCRIPTION}
  \item{\code{RMETRCOD}}{integer COLUMN_DESCRIPTION}
  \item{\code{AGLCOD}}{integer COLUMN_DESCRIPTION}
  \item{\code{ANOINST}}{integer COLUMN_DESCRIPTION}
  \item{\code{ANOEXT}}{integer COLUMN_DESCRIPTION}
  \item{\code{SUCESSOR}}{integer COLUMN_DESCRIPTION}
  \item{\code{LATITUDE}}{double COLUMN_DESCRIPTION}
  \item{\code{LONGITUDE}}{double COLUMN_DESCRIPTION}
  \item{\code{ALTITUDE}}{integer COLUMN_DESCRIPTION}
  \item{\code{AREA}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{ftp://ftp.datasus.gov.br/dissemin/publicos/SINASC/1996_/TABELAS/CADMUN.DBF}
}
\usage{
CADMUN
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
