% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CID10}
\alias{CID10}
\title{10th International Classification of Diseases}
\format{
A data frame with 14198 rows and 6 variables:
\describe{
  \item{\code{CID10}}{integer COLUMN_DESCRIPTION}
  \item{\code{OPC}}{integer COLUMN_DESCRIPTION}
  \item{\code{CAT}}{integer COLUMN_DESCRIPTION}
  \item{\code{SUBCAT}}{integer COLUMN_DESCRIPTION}
  \item{\code{DESCR}}{integer COLUMN_DESCRIPTION}
  \item{\code{RESTRSEXO}}{integer COLUMN_DESCRIPTION}
}
}
\source{
\url{ftp://ftp.datasus.gov.br/dissemin/publicos/SINASC/1996_/TABELAS/CID10.DBF}
}
\usage{
CID10
}
\description{
The 10th International Classification of Diseases (ICD-10) is a global system for classifying diseases,
health conditions, and a wide range of health-related events. Developed by the World Health Organization (WHO), it serves as a standard
diagnostic tool for epidemiology, clinical purposes, and health management. ICD-10 is used worldwide for health reporting, statistical
analysis, and medical billing.
}
\keyword{datasets}
