% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.sinasc.R
\name{label.sinasc}
\alias{label.sinasc}
\title{Insert labels in raw SINASC data}
\usage{
label.sinasc(data)
}
\arguments{
\item{data}{Raw SINASC data generated by the function download.sinasc with all columns.}
}
\value{
A data frame from SINASC with all codes replaced by their respective labels.
}
\description{
The function download.sinasc create a data frame about the SINASC data with no treatment. This function is usual to realize the
treatment of the data and replace the codes by respective labels.
}
\seealso{
\code{\link{download.sinasc} \link{panel.sinasc}}
}
\author{
Luan Augusto, \email{luanguto87@gmail.com}
}
