% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ocup}
\alias{ocup}
\title{CBO 2002 Estructure}
\format{
A data frame with 2694 rows and 2 variables:
\describe{
  \item{\code{CODIGO}}{integer CBO 2002 code}
  \item{\code{TITULO}}{character Title of the occupation}
}
}
\source{
\url{https://www.gov.br/trabalho-e-emprego/pt-br/assuntos/cbo/servicos/downloads/estrutura-cbo.zip}
}
\usage{
ocup
}
\description{
Dataset with the CBO 2002 code and their respectives titles.
}
\keyword{datasets}
