% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.sinasc.R
\name{panel.sinasc}
\alias{panel.sinasc}
\title{Panel SINASC}
\usage{
panel.sinasc(x)
}
\arguments{
\item{x}{SINASC data frame generated by the function label.sinasc.}
}
\value{
A data frame from SINASC with the columns needed for the Power BI dashboard.
}
\description{

}
\seealso{
\code{\link{download.sinasc} \link{label.sinasc}}
}
\author{
Luan Augusto, \email{luanguto87@gmail.com}
}
