% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TABUF}
\alias{TABUF}
\title{UF Structure}
\format{
A data frame with 27 rows and 3 variables:
\describe{
  \item{\code{SIGLA_UF}}{integer COLUMN_DESCRIPTION}
  \item{\code{CODIGO}}{integer COLUMN_DESCRIPTION}
  \item{\code{DESCRICAO}}{integer COLUMN_DESCRIPTION}
}
}
\source{
\url{ftp://ftp.datasus.gov.br/dissemin/publicos/SINASC/1996_/TABELAS/TABUF.DBF}
}
\usage{
TABUF
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
