% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.sinasc.R
\name{panel.sinasc}
\alias{panel.sinasc}
\title{Panel SINASC}
\usage{
panel.sinasc(data)
}
\arguments{
\item{data}{SINASC data frame generated by the function label.sinasc.}
}
\value{
A data frame from SINASC with the columns needed for the Power BI dashboard.
}
\description{
The function label.sinasc treat the data frame about the SINASC after download with the download.sinasc function.
But many columns are unnecessary for the Power BI Dashboard, so this function select the columns needed and reat the columns names.
}
\seealso{
\code{\link{download.sinasc} \link{process.sinasc}}
}
\author{
Luan Augusto, \email{luanguto87@gmail.com}
}
